+++ adminforth/plugins/email-password-reset/custom/ResetPasswordUnderLogin.vue
@@ -1,8 +1,8 @@
 <template>
   <p class="text-gray-500 dark:text-gray-400 font-sm text-left mt-3">
-    Forgot password? 
+    {{$t('Forgot password?')}} 
     <Link to="/reset-password">
-      reset it
+      {{$t('reset it')}}
     </Link>
   </p>
 </template>