+++ adminforth/plugins/i18n/custom/LanguageUnderLogin.vue
@@ -4,7 +4,6 @@
       v-model="selectedLanguage"
       :options="options"
       :placeholder="$t('Select language')"
-      @change="changeLanguage"
     >
       <template #item="{ option }">
         <span class="mr-1">
@@ -31,7 +30,7 @@
 <script setup>
 import Select from '@/afcl/Select.vue';
 import 'flag-icon-css/css/flag-icons.min.css';
-import { setLang } from './langCommon';
+import { setLang, getCountryCodeFromLangCode, getLocalLang } from './langCommon';
 
 import { computed, ref, onMounted, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
@@ -43,33 +42,10 @@ const props = defineProps(['meta', 'resource']);
 
 const selectedLanguage = ref('');
 
-
-
 watch(() => selectedLanguage.value, (newVal) => {
-  localStorage.setItem(LS_LANG_KEY, newVal);
-  
- 
-
   setLang({ setLocaleMessage, locale }, props.meta.pluginInstanceId, newVal);
 });
 
-
-// only remap the country code for the languages where language code is different from the country code
-// don't include es: es, fr: fr, etc, only include the ones where language code is different from the country code
-const countryISO31661ByLangISO6391 = {
-    en: 'us', // English → United States
-    zh: 'cn', // Chinese → China
-    hi: 'in', // Hindi → India
-    ar: 'sa', // Arabic → Saudi Arabia
-    ko: 'kr', // Korean → South Korea
-    ja: 'jp', // Japanese → Japan
-    uk: 'ua', // Ukrainian → Ukraine
-};
-
-function getCountryCodeFromLangCode(langCode) {
-    return countryISO31661ByLangISO6391[langCode] || langCode;
-}
-  
 
 const options = computed(() => {
   return props.meta.supportedLanguages.map((lang) => {
@@ -80,11 +56,17 @@ const options = computed(() => {
   });
 });
 
-const LS_LANG_KEY = `${props.meta.brandSlug}-lang`;
-
 onMounted(() => {
   console.log('LanguageUnderLogin mounted', props.meta.supportedLanguages);
-  selectedLanguage.value = localStorage.getItem(LS_LANG_KEY) || props.meta.supportedLanguages[0].code;
+  selectedLanguage.value = getLocalLang(props.meta.supportedLanguages);
+  setLang({ setLocaleMessage, locale }, props.meta.pluginInstanceId, selectedLanguage.value);
+  // todo this mounted executed only on this component mount, f5 from another page apart login will not read it
 });
+
+
+
+
+
+
 
 </script>