+++ adminforth/servers/express.ts
@@ -247,6 +247,15 @@ class ExpressServer implements IExpressHttpServer {
     };
   }
 
+  translatable(handler) {
+    // same as authorize, but injects tr function into request
+    return async (req, res, next) => {
+      const tr = (msg: string, category: string, params: any): Promise<string> => this.adminforth.tr(msg, category, req.headers['accept-language'], params);
+      req.tr = tr;
+      handler(req, res, next);
+    }
+  }
+
   endpoint({ method='GET', path, handler, noAuth=false }) {
     if (!path.startsWith('/')) {
       throw new Error(`Path must start with /, got: ${path}`);
@@ -290,7 +299,10 @@ class ExpressServer implements IExpressHttpServer {
         }
         
       };
-      const input = { body, query, headers, cookies, adminUser, response, _raw_express_req: req, _raw_express_res: res};
+
+      const acceptLang = headers['accept-language'];
+      const tr = (msg: string, category: string, params: any): Promise<string> => this.adminforth.tr(msg, category, acceptLang, params);
+      const input = { body, query, headers, cookies, adminUser, response, _raw_express_req: req, _raw_express_res: res, tr};
       
       let output;
       try {