+++ dev-demo/resources/translation.ts
@@ -1,39 +1,42 @@
 import AdminForth, {
   AdminForthDataTypes,
-  AdminForthResource,
-  AdminForthResourceColumn,
   AdminForthResourceInput,
-  AdminUser,
 } from "../../adminforth";
 import CompletionAdapterOpenAIChatGPT from "../../adminforth/adapters/completion-adapter-open-ai-chat-gpt";
 import I18nPlugin from "../../adminforth/plugins/i18n";
 import { v1 as uuid } from "uuid";
 
 
-// model translations {
-//   en_string   @id
-//   created_at  DateTime
-//   uk_string   String
-//   jp_string   String
-//   fr_string   String
-// }
 export default {
   dataSource: "maindb",
   table: "translations",
   resourceId: "translations",
   label: "Translations",
 
-  recordLabel: (r: any) => `👤 ${r.email}`,
+  recordLabel: (r: any) => `👤 ${r.en_string}`,
   plugins: [
     new I18nPlugin({
       supportedLanguages: ['en', 'uk', 'ja', 'fr'],
+
+      // names of the fields in the resource which will store translations
       translationFieldNames: {
         en: 'en_string',
         uk: 'uk_string',
         ja: 'ja_string',
         fr: 'fr_string',
       },
+
+      // name of the field which will store the category of the string
+      // this helps to categorize strings and deliver them efficiently
+      categoryFieldName: 'category',
+
+      // optional field to store the source (e.g. source file name)
       sourceFieldName: 'source',
+
+      // optional field store list of completed translations
+      // will hel to filter out incomplete translations
+      completedFieldName: 'completedLangs',
+
       completeAdapter: new CompletionAdapterOpenAIChatGPT({
         openAiApiKey: process.env.OPENAI_API_KEY as string,
       }),
@@ -44,11 +47,18 @@ export default {
     allowedActions: {
     
     },
+    listPageSize: 30,
   },
   columns: [
     {
-      name: "en_string",
+      name: "id",
+      fillOnCreate: ({ initialRecord, adminUser }: any) => uuid(),
       primaryKey: true,
+      showIn: [],
+    },
+    {
+      name: "en_string",
+      type: AdminForthDataTypes.STRING,
     },
     {
       name: "created_at",
@@ -56,20 +66,28 @@ export default {
     },
     {
       name: "uk_string",
+      type: AdminForthDataTypes.STRING,
     },
     {
       name: "ja_string",
+      type: AdminForthDataTypes.STRING,
     },
     {
       name: "fr_string",
+      type: AdminForthDataTypes.STRING,
+    },
+    {
+      name: "completedLangs",
     },
     {
       name: "source",
+      showIn: ['filter', 'show'],
+      type: AdminForthDataTypes.STRING,
+    },
+    {
+      name: "category",
+      showIn: ['filter', 'show', 'list'],
+      type: AdminForthDataTypes.STRING,
     }
-
-
   ],
-  hooks: {
-    
-  },
 } as AdminForthResourceInput;